hello

+++ app/Http/Controllers/LikeController.php
@@ -15,16 +15,32 @@ class LikeController extends Controller
         $data = [
             'user_id' => auth()->user()->id,
             'posts_id' => $post->id,
-            'status' => request()->status
         ];
         // dd($data);
         if(Like::where($data)->exists())
         {
-            Like::where($data)->delete();
+            // dd('hello');
+            // dd(Like::where($data)->where('status',0)->exists());
+            $like = Like::Where($data)->get();
+            if($like->first()->status == request()->status)
+            {
+                Like::where($data)->delete();
+            }
+            else
+            {
+                Like::Where($data)->update(['status' => request()->status]);
+            }
         }
         else
         {
-            Like::create($data);
+            $data['status'] = request()->status;
+            $like = Like::create($data);
+            auth()->user()->likes()->attach($like);
         }
+
+        $like = Like::Where('status',1)->Where('posts_id',$post->id)->count();
+        $dislike = Like::Where('status',0)->Where('posts_id',$post->id)->count();
+        return compact(['like','dislike']);
+
     }
 }
